Remove the task output queue.
It was a kind of crummy abstraction, so now each worker can report
progress whenever it wants with arbitrary json attached. The controller
will hang on to this data for the user. As part of this, took some steps
to cleaning up the way the controller stores job metadata to be more
flexible and threadsafe.